# docs/api-contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "ApiErrorPayload": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorPayload",
        "type": "object"
      },
...
      "ApiResponse_CustomerData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[CustomerData]",
        "type": "object"
      },
...
      "ApiResponse_SelfIdentityData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelfIdentityData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[SelfIdentityData]",
        "type": "object"
      },
...
      "CustomerData": {
        "properties": {
          "customer_id": {
            "default": "",
            "title": "Customer Id",
            "type": "string"
          },
          "display_name": {
            "default": "",
            "title": "Display Name",
            "type": "string"
          },
          "last_contact_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Contact At"
          },
          "remark": {
            "default": "",
            "title": "Remark",
            "type": "string"
          },
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "title": "CustomerData",
        "type": "object"
      },
      "CustomerPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "remark": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "status": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "CustomerPatchRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SelfIdentityData": {
        "properties": {
          "display_name": {
            "default": "",
            "title": "Display Name",
            "type": "string"
          },
          "identity_facts": {
            "items": {
              "type": "string"
            },
            "title": "Identity Facts",
            "type": "array"
          }
        },
        "title": "SelfIdentityData",
        "type": "object"
      },
      "SelfIdentityGenerateRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "SelfIdentityGenerateRequest",
        "type": "object"
      },
      "SelfIdentityPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "identity_facts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identity Facts"
          }
        },
        "title": "SelfIdentityPatchRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WeChat AI Local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/customers/{customer_id}": {
      "get": {
        "operationId": "get_customer_api_v1_customers__customer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CustomerData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Customer",
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "operationId": "update_customer_api_v1_customers__customer_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CustomerData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Customer",
        "tags": [
          "customers"
        ]
      }
    },
...
    "/api/v1/identity/self/global": {
      "get": {
        "operationId": "get_global_self_identity_api_v1_identity_self_global_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SelfIdentityData_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Global Self Identity",
        "tags": [
          "identity"
        ]
      },
      "patch": {
        "operationId": "update_global_self_identity_api_v1_identity_self_global_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfIdentityPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SelfIdentityData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Global Self Identity",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/identity/self/global/generate": {
      "post": {
        "operationId": "generate_global_self_identity_api_v1_identity_self_global_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfIdentityGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SelfIdentityData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Global Self Identity",
        "tags": [
          "identity"
        ]
      }
    },
...
  }
}